requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
python-telegram-bot>=21.0
psycopg2-binary>=2.9.0

# Avito Dashboard
fastapi>=0.109.0
uvicorn>=0.27.0
openpyxl>=3.1.0
python-multipart>=0.0.6